Basic stamina system in place
To note: This is only for sprinting. Attacking will have stamina degradation later. Currently stamina does not deplete if sprinting while attacking, but that will be remedied. Also keep an eye on sprint speed going into crawl when too many inputs are received; saw it happen once but was unable to replicate.