Roll, stamina fix/update. Removed unused notifies
Proper stamina cost added for roll, with checks. Checks added for other actions. Stamina maximum added for check instead of flat value so player can adjust it, when stat leveling is added in the future. Notify added for roll, for future animation adjustment. Current animation does not cover desirable distance, so it either needs to be adjusted or completely replaced in the future. Current works as a placeholder