Refined Lock On system
I might test this further, but the Lock On system is near a finalized state, if not already at it.
Other options are listed in the comments, if I wish to make it more complex. However, it cycles through a group of three enemies just fine, and I currently don't have plans to have encounters contain more than that. Perhaps later, but I don't want this game to become musou-like, which is why I hesitate.