Testing custom environment sounds/lighting
Created a few custom sounds for use in this project. Testing out one within a triggerbox in the cave level.
Adjusted lighting of character's lantern, making the radius closer to the source stronger while having a definitive, non-expansive space. Light will still creep away from the source, but there is a spot around the character that should clearly create an area around the player, now.
Removed a few test objects from lock-on tests.