Added UI/audio to load zones, to be more intuitive
When a player is near a load zone, they are gently informed through audio/visual cues beforehand, if they wish to proceed. Included ambient sound for said audio cue. It may be edited in the future to sound less like a forest breeze and more like a stray gust into a cavern.
Minor refactoring of load zone resources.