Testing custom texture import onto glb model
Custom textures can be imported as glb/glTF file types, NOT fbx like most are used to.
There's some roughness, notably with lighting and seams/edges of model, but the vast majority of the texture is present!
In the future I may use fbx for basic models that are textured simply, but glb/glTF export seems to have what I need.